modo sin conexion usable por todas las vistas pero sin BD de productos

+++ www/js/config.js
@@ -21,6 +21,30 @@ function set_settings(url_server){
 // hace todos los ping consecutivamente
 function ping(url){
 
+  read_alerts();
+
+  if ($('#alert').text().includes("Modo sin conexion")){
+    try {
+      modo_sin_conexion();
+    }
+    catch(err) {
+      // "no se soporta un modo sin conexion, ej en login & register"
+    }
+
+    $('.alert.alert-danger button').remove()
+    $('.alert.alert-danger').append('\
+      <button type="button" class="close" onclick="location.reload();">\
+        <span class="fa fa-refresh" aria-hidden="true"></span>\
+      </button>')
+    dis_path = window.location.pathname.split("/").pop()
+    if (dis_path == "login.html" || dis_path=="crear_perfil.html"){
+      $('.alert b').text("");
+      $('.alert b').append("ERROR: Necesitas conexion a internet");
+      $('.alert').css("margin-bottom","0");
+    }
+    return;
+  }
+
   new_url = all_servers.splice(0,1)[0];
   if (new_url!=undefined && url_server=="") {
     setTimeout( function(){
@@ -36,7 +60,7 @@ function ping(url){
         pings.done(function (response) {
           url_server = url;
           console.log("Conected to", url_server);
-          read_alerts();
+          // read_alerts();
           try {
             get_my_data();
           }
@@ -46,25 +70,33 @@ function ping(url){
         })
         pings.fail(function (response) {
           console.log("fail", url);
+          // read_alerts();
           if (new_url == undefined && url_server==""){
-            read_alerts();
-            send_alert('\
-              <!--<span style="font-size: 10px">\
-                Problema de conexion con el servidor\
-              </span><br> -->\
-              <b>Modo sin conexion activado</b>', "danger");
-            // solo se recaarga la pagina si no esta el mensaje de alerta de sin conexion
-            if ($('#alert').text() == ""){
+            // if ($('#alert').text() == ""){ //se recarga la pagina si no hay conexion a internet y no se ha enviado la alerta
+              send_alert('<b>Modo sin conexion activado</b>', "danger");
               location.reload();
-            }
-            else{
-              // console.log("$('#alert').text() != ''")
-              $('.alert.alert-danger button').remove()
-              $('.alert.alert-danger').append('\
-                <button type="button" class="close" onclick="location.reload();">\
-                  <span class="fa fa-refresh" aria-hidden="true"></span>\
-                </button>')
-            }
+            // }
+            // else{ //si el modo sin conexion esta activado (y el mensaje se esta mostrando)
+
+            //   try {
+            //     modo_sin_conexion();
+            //   }
+            //   catch(err) {
+            //     // "no se soporta un modo sin conexion"
+            //   }
+
+            //   $('.alert.alert-danger button').remove()
+            //   $('.alert.alert-danger').append('\
+            //     <button type="button" class="close" onclick="location.reload();">\
+            //       <span class="fa fa-refresh" aria-hidden="true"></span>\
+            //     </button>')
+            //   dis_path = window.location.pathname.split("/").pop()
+            //   if (dis_path == "login.html" || dis_path=="crear_perfil.html"){
+            //     $('.alert b').text("");
+            //     $('.alert b').append("ERROR: Necesitas conexion a internet");
+            //     $('.alert').css("margin-bottom","0");
+            //   }
+            // }
           }
         })
       }
@@ -75,7 +107,7 @@ function ping(url){
     pings.done(function (response) {
       url_server = url;
       console.log("Conected to", url_server);
-      read_alerts();
+      // read_alerts();
       try {
         get_my_data();
       }
@@ -83,22 +115,50 @@ function ping(url){
         // "no se requiere un get_my_data"
       }
     })
-    // pings.fail(function (response) {
-    //   console.log("fail", url);
-    //   if (new_url == undefined && url_server==""){
-    //     alert("Problema de conexion con el servidor");
-    //     location.reload();
-    //   }
-    // })
   }
-
 };
+
+// para poder almacenar imagenes con local storage (la foto de perfil del usuario)
+function toDataUrl(url, callback) {
+  var xhr = new XMLHttpRequest();
+  xhr.responseType = 'blob';
+  xhr.onload = function() {
+    var reader = new FileReader();
+    reader.onloadend = function() {
+      callback(reader.result);
+    }
+    reader.readAsDataURL(xhr.response);
+  };
+  xhr.open('GET', url);
+  xhr.send();
+}
+// se usa:
+// toDataUrl('http://example/url', function(base64Img) {
+//   console.log(base64Img);
+// });
+
+
 
 
 // el mensaje debe venir con el formato que se quiere mostrar... status debe ser: success, info, warning o danger
 function send_alert(message, status){
-  // Guardar mensaje
-  localStorage.setItem('alert_data', JSON.stringify({ 'alert_message': message, 'alert_status': status }));
+  var exist_alert = localStorage.getItem('alert_data')
+  if (exist_alert){
+    append_on_alert(message);
+  }
+  else{
+    // Guardar mensaje
+    localStorage.setItem('alert_data', JSON.stringify({ 'alert_message': message, 'alert_status': status }));
+  }
+}
+
+// para que se pueda agregar algo a una alerta ya existente
+function append_on_alert(message){
+  
+  alert = JSON.parse(localStorage.getItem('alert_data'));
+
+  localStorage.setItem('alert_data', JSON.stringify({ 'alert_message': alert.alert_message+'<br>'+message, 'alert_status': status }));
+
 }
 
 function read_alerts(){
@@ -115,7 +175,7 @@ function read_alerts(){
     localStorage.removeItem('alert_data');
 
     $("#alert").html('\
-      <div class="alert alert-'+status+' alert-dismissible fade in" role="alert" style="margin-bottom: 0;">\
+      <div class="alert alert-'+status+' alert-dismissible fade in" role="alert">\
         <button type="button" class="close" data-dismiss="alert" aria-label="Close">\
           <span aria-hidden="true">&times;</span>\
         </button>\
